% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_for_overlaps.R
\name{test_for_overlaps}
\alias{test_for_overlaps}
\title{Test for Overlaps}
\usage{
test_for_overlaps(
  matched_aavenger_sites_df,
  list_of_Rdata_files,
  overlap_ranges_to_test
)
}
\arguments{
\item{matched_aavenger_sites_df}{A \code{data.frame} object containing specifically formatted match and insertion sites for each heatmap_group. See \code{format_aavenger_sites()}.}

\item{list_of_Rdata_files}{A \code{list} object of RData files containing a \code{GRanges} object called \code{epigenData} with the following columns:
\describe{
\item{seqname}{name of the chromosome. Must match those in \code{matched_aavenger_sites_df}}.
\item{ranges}{An \code{IRanges} object with a genomic range.}
\item{Rle}{The strand the range is located in. Should be set to \code{*}.}
\item{mid}{The mid point between the two values in the \code{ranges} column.}}}

\item{overlap_ranges_to_test}{A \code{numeric} \code{vector} of the different genomic windows to use in the overlap test. Values must be integers.}
}
\value{
A dataframe containing the original \code{matched_aavenger_sites_df} and a column of overlap counts per each file from \code{list_of_Rdata_files}.
}
\description{
Test a list of Rdata files for genomic overlaps with integration site positions and count the number of overlaps.
}
\examples{
test_for_overlaps(matched_aavenger_sites_df, list_of_Rdata_files = list.files(path_to_RData_files), overlap_ranges_to_test = c(1000,10000,1000000))

}
